Treat all preformatted blocks as code
Some code examples are not tagged with a specific programming language,
which is okay from a content perspective. However the CSS relied on it
being tagged in order to apply styling and some code was not properly
formatted.

We loose some specificity with this fix. This being mostly a tech blog,
I assume we won't be writing much else than code in <pre> blocks and
even then it doesn't look bad.