Always specify page language
Specifying lanuage is important for assistive technologies like screen
readers.

The overall language of the blog is English, but there are some posts in
German.

I had initially omitted the language on the posts index, because I did
not know how to represent multiple languages. However I found out the
lang attribute is not specific to the <html> element.

Reference:

https://html.spec.whatwg.org/multipage/dom.html#the-lang-and-xml:lang-attributes