Always show the date and author below the post title
The previous iteration was closer to how the metadata was displayed on
the old Wordpress blog.

I'd argue that grouping the author and date with the title makes it
easier to find these details. Seeing the date right away will probably
inform the decision whether you want to read a post.