Work around image processing error
The build fails with the following error when using newer eleventy-img
versions (could be due to the simultaneous upgrade of one of its
dependencies):

    VipsOperation: class "magicksave" not found

The image in question was the only GIF on the blog. As it's a static
image it can also be encoded as a PNG, which is handled correctly.