Add post about backlog maintenance (#90)
Every once and again, I see projects with really big backlogs. I wanted to
articulate why I think that's not a good thing and what can be done about it.

Initially the post also contained a point about not tracking technical debt in
the backlog. That triggered a longer discussion during review. Addressing the
nuances of handling technical debt would be too distracting for a general post
about backlogs, so I removed that part altogether. It's something we can write
about in a dedicated blog post.