Identify posts collection by path
We intend to add tags to each post. That will be achieved with the
`tags` key in each post's frontmatter.

The `tags` key is already used in the posts directory frontmatter
(posts.json) to mark the collection of all posts. In 11ty, the data from
the frontmatter in individual posts takes precendence over the
directory-level data. Because of this the `post` tag will be overwritten
for posts that define tags in their frontmatter and these posts will no
longer be in the `post` collection. (We use this collection to list
posts on the index page)

Luckily there are other ways to define collections, like the Eleventy
collection API used in this commit.

Documentation:

- Data cascade: https://www.11ty.dev/docs/data-cascade/
- Collections: https://www.11ty.dev/docs/collections/